Refactored out a form view.
